ISSUE #682 split SupermeshNode type into its own file
