+++ bouncer/src/repo/core/model/bson/repo_node_mesh.h
@@ -222,42 +222,6 @@ namespace repo {
 				*/
 				std::vector<uint32_t> getFacesSerialized() const;
 			};
-
-			class REPO_API_EXPORT SupermeshNode : public MeshNode
-			{
-			public:
-
-				SupermeshNode()
-				{
-				}
-
-				SupermeshNode(RepoBSON bson,
-					const std::unordered_map<std::string, std::pair<std::string, std::vector<uint8_t>>>& binMapping) : MeshNode(bson, binMapping)
-				{
-				}
-
-				std::vector<repo_mesh_mapping_t> getMeshMapping() const;
-				std::vector<float> getSubmeshIds() const;
-
-				SupermeshNode cloneAndUpdateIds(
-					repo::lib::RepoUUID uniqueID,
-					repo::lib::RepoUUID sharedID);
-
-				/**
-				* Create a new copy of the node and update its mesh mapping
-				* @return returns a new meshNode with the new mappings
-				*/
-				SupermeshNode cloneAndUpdateMeshMapping(
-					const std::vector<repo_mesh_mapping_t>& vec,
-					const bool& overwrite = false);
-
-				/**
-				* Given a mesh mapping, convert it into a bson object
-				* @param mapping the mapping to convert
-				* @return return a bson object containing the mapping
-				*/
-				static RepoBSON meshMappingAsBSON(const repo_mesh_mapping_t& mapping);
-			};
 		} //namespace model
 	} //namespace core
 } //namespace repo
